Crate rusoto_meteringmarketplace

AWS Marketplace Metering Service
This reference provides descriptions of the low-level AWS Marketplace Metering Service API.

AWS Marketplace sellers can use this API to submit usage data for custom usage dimensions.

Submitting Metering Records

MeterUsage- Submits the metering record for a Marketplace product. MeterUsage is called from an EC2 instance.

BatchMeterUsage- Submits the metering record for a set of customers. BatchMeterUsage is called from a software-as-a-service (SaaS) application.

Accepting New Customers

ResolveCustomer- Called by a SaaS application during the registration process. When a buyer visits your website during the registration process, the buyer submits a Registration Token through the browser. The Registration Token is resolved through this API to obtain a CustomerIdentifier and Product Code.

Entitlement and Metering for Paid Container Products

Paid container software products sold through AWS Marketplace must integrate with the AWS Marketplace Metering Service and call the RegisterUsage operation for software entitlement and metering. Calling RegisterUsage from containers running outside of Amazon Elastic Container Service (Amazon ECR) isn't supported. Free and BYOL products for ECS aren't required to call RegisterUsage, but you can do so if you want to receive usage data in your seller reports. For more information on using the RegisterUsage operation, see Container-Based Products.

BatchMeterUsage API calls are captured by AWS CloudTrail. You can use Cloudtrail to verify that the SaaS metering records that you sent are accurate by searching for records with the eventName of BatchMeterUsage. You can also use CloudTrail to audit records over time. For more information, see the AWS CloudTrail User Guide .

If you're using the service, you're probably looking for MarketplaceMeteringClient and MarketplaceMetering.

Structs

BatchMeterUsageRequest
A BatchMeterUsageRequest contains UsageRecords, which indicate quantities of usage within your application.

BatchMeterUsageResult
Contains the UsageRecords processed by BatchMeterUsage and any records that have failed due to transient error.

MarketplaceMeteringClient
A client for the AWSMarketplace Metering API.

MeterUsageRequest

MeterUsageResult

RegisterUsageRequest

RegisterUsageResult

ResolveCustomerRequest
Contains input to the ResolveCustomer operation.

ResolveCustomerResult
The result of the ResolveCustomer operation. Contains the CustomerIdentifier and product code.

UsageRecord
A UsageRecord indicates a quantity of usage for a given product, customer, dimension and time.

Multiple requests with the same UsageRecords as input will be deduplicated to prevent double charges.

UsageRecordResult
A UsageRecordResult indicates the status of a given UsageRecord processed by BatchMeterUsage.

Enums

BatchMeterUsageError
Errors returned by BatchMeterUsage

MeterUsageError
Errors returned by MeterUsage

RegisterUsageError
Errors returned by RegisterUsage

ResolveCustomerError
Errors returned by ResolveCustomer

Traits

MarketplaceMetering
Trait representing the capabilities of the AWSMarketplace Metering API. AWSMarketplace Metering clients implement this trait.