Struct rusoto_meteringmarketplace::MarketplaceMeteringClient

pub struct MarketplaceMeteringClient { /* fields omitted */ }

A client for the AWSMarketplace Metering API.

Methods

impl MarketplaceMeteringClient

pub fn new(region: Region) -> MarketplaceMeteringClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> MarketplaceMeteringClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,